requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.3.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
cryptography>=41.0.0
pycryptodome>=3.18.0
web3>=6.10.0
eth-account>=0.9.0
py-solc-x>=1.12.0
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0 